TASK-81.2 Commit B: extract core::settings
Lift the entire settings layer (schema + JSON IO + lock-free caches)
from the Tauri shell into core/. ~700 LOC of platform-agnostic state
that the original shell module owned simply because it grew there
first.

- core::settings: schema types (HotkeyKind/Config/Settings/Target,
  AudioSettings, BehaviorSettings, PillSettings, StatsSettings),
  envelope IO over &Path, and every cache static. The three behavior
  AtomicBool/AtomicU64 mirrors that lived in shell behavior.rs
  (SHOW_IN_FULLSCREEN, PAUSE_AUDIO, BT_RESUME_DELAY_MS) move alongside
  their schema. BT_RESUME_DELAY_MS_MAX = 10000 const replaces the
  inline `.min(10_000)` clamp.
- core::settings now owns serde_json (new dep) for the on-disk
  envelope read/write. dirs / config-dir resolution stays out of
  core — shell passes a &Path on every load_*/save_*.
- Tauri `apps/tauri/src-tauri/src/settings/mod.rs` becomes a thin
  facade: resolves Tauri's app_config_dir(), re-exports the schema
  types other shell modules import via crate::settings::*, and hosts
  the #[tauri::command] wrappers.
- behavior.rs drops the atomic statics and accessors; keeps
  apply_collection_behavior (NSPanel platform glue) and the four
  behavior_* commands as 1-liner-ish wrappers over core setters.
- lib.rs setup hydration + hot-path readers point at
  openwhisper_core::settings::* directly. media_control/windows.rs
  reads bt_resume_delay_ms() from core.

Coverage: 8 schema tests move with the code (legacy JSON migration,
clamp, defaults). cargo check clean across default, tauri, and
macos-shell. cargo test green (50 prior + 8 settings = 58 core,
0 tauri lib — 8 migrated up). Playwright 87/87.
