+++ core/src/lib.rs
@@ -4,6 +4,7 @@ mod ffi_c;
 pub mod media_gate;
 #[cfg(feature = "recognizer")]
 pub mod recognizer;
+pub mod settings;
 pub mod stats;
 pub mod store;
 pub mod transcript;
